Server correctly lets unassigned players leave game
Because the getColorOf function in the Game class didn't find the color of the
unassigned player, the leaveGame function was incorrectly bailing early for
players who joined the game without being assigned a color. So users would see
an open game even after the unassigned player had disconnected. Fixed this.

Also, set getColorOf, getOpponent and getActiveOpponent methods on Game to
return undefined instead of null for nonexistent players, just for a bit more
clarity.